# Conflicts:
#	src/main/java/in/koreatech/koin/global/domain/notification/service/NotificationService.java
#	src/test/java/in/koreatech/koin/AcceptanceTest.java

+++ src/main/java/in/koreatech/koin/global/domain/upload/controller/UploadApi.java
@@ -59,7 +59,7 @@ public interface UploadApi {
     ResponseEntity<UploadUrlResponse> getPresignedUrl(
         @PathVariable ImageUploadDomain domain,
         @RequestBody @Valid UploadUrlRequest request,
-        @Auth(permit = {OWNER, STUDENT, COOP}, anonymous = true) Long memberId
+        @Auth(permit = {OWNER, STUDENT, COOP}, anonymous = true) Integer memberId
     );
 
     @ApiResponses(
@@ -90,7 +90,7 @@ public interface UploadApi {
     ResponseEntity<UploadFileResponse> uploadFile(
         @Parameter(in = PATH) @PathVariable ImageUploadDomain domain,
         @RequestPart MultipartFile multipartFile,
-        @Auth(permit = {OWNER, STUDENT, COOP}, anonymous = true) Long memberId
+        @Auth(permit = {OWNER, STUDENT, COOP}, anonymous = true) Integer memberId
     );
 
     @ApiResponses(
@@ -121,6 +121,6 @@ public interface UploadApi {
     ResponseEntity<UploadFilesResponse> uploadFiles(
         @Parameter(in = PATH) @PathVariable ImageUploadDomain domain,
         @RequestPart List<MultipartFile> files,
-        @Auth(permit = {OWNER, STUDENT, COOP}, anonymous = true) Long memberId
+        @Auth(permit = {OWNER, STUDENT, COOP}, anonymous = true) Integer memberId
     );
 }
